fix: teleport controlextension timeouts
