+++ packages/core/mesh/teleport/src/teleport.ts
@@ -4,24 +4,15 @@
 
 import { type Duplex } from 'node:stream';
 
-import {
-  asyncTimeout,
-  scheduleTaskInterval,
-  runInContextAsync,
-  synchronized,
-  scheduleTask,
-  type Event,
-} from '@dxos/async';
+import { runInContextAsync, synchronized, scheduleTask, type Event } from '@dxos/async';
 import { Context } from '@dxos/context';
 import { failUndefined } from '@dxos/debug';
 import { invariant } from '@dxos/invariant';
 import { PublicKey } from '@dxos/keys';
 import { log } from '@dxos/log';
-import { schema, RpcClosedError, TimeoutError } from '@dxos/protocols';
-import { type ControlService } from '@dxos/protocols/proto/dxos/mesh/teleport/control';
-import { createProtoRpcPeer, type ProtoRpcPeer } from '@dxos/rpc';
-import { Callback } from '@dxos/util';
+import { RpcClosedError, TimeoutError } from '@dxos/protocols';
 
+import { ControlExtension } from './control-extension';
 import { type CreateChannelOpts, Muxer, type MuxerStats, type RpcPort } from './muxing';
 
 export type TeleportParams = {
@@ -51,18 +42,7 @@ export class Teleport {
 
   private readonly _muxer = new Muxer();
 
-  private readonly _control = new ControlExtension({
-    heartbeatInterval: CONTROL_HEARTBEAT_INTERVAL,
-    heartbeatTimeout: CONTROL_HEARTBEAT_TIMEOUT,
-    onTimeout: () => {
-      if (this._destroying || this._aborting) {
-        return;
-      }
-      // TODO(egorgripasov): Evaluate use of abort instead of destroy.
-      log('abort teleport due to onTimeout in ControlExtension');
-      this.abort(new TimeoutError('control extension')).catch((err) => log.catch(err));
-    },
-  });
+  private readonly _control;
 
   private readonly _extensions = new Map<string, TeleportExtension>();
   private readonly _remoteExtensions = new Set<string>();
@@ -78,6 +58,22 @@ export class Teleport {
     this.initiator = initiator;
     this.localPeerId = localPeerId;
     this.remotePeerId = remotePeerId;
+
+    this._control = new ControlExtension(
+      {
+        heartbeatInterval: CONTROL_HEARTBEAT_INTERVAL,
+        heartbeatTimeout: CONTROL_HEARTBEAT_TIMEOUT,
+        onTimeout: () => {
+          if (this._destroying || this._aborting) {
+            return;
+          }
+          log.info('abort teleport due to onTimeout in ControlExtension');
+          this.abort(new TimeoutError('control extension')).catch((err) => log.catch(err));
+        },
+      },
+      this.localPeerId,
+      this.remotePeerId,
+    );
 
     this._control.onExtensionRegistered.set(async (name) => {
       log('remote extension', { name });
@@ -272,85 +268,4 @@ export interface TeleportExtension {
   onOpen(context: ExtensionContext): Promise<void>;
   onClose(err?: Error): Promise<void>;
   onAbort(err?: Error): Promise<void>;
-}
-
-type ControlRpcBundle = {
-  Control: ControlService;
-};
-
-type ControlExtensionOpts = {
-  heartbeatInterval: number;
-  heartbeatTimeout: number;
-  onTimeout: () => void;
-};
-
-class ControlExtension implements TeleportExtension {
-  private readonly _ctx = new Context({
-    onError: (err) => {
-      this._extensionContext.close(err);
-    },
-  });
-
-  public readonly onExtensionRegistered = new Callback<(extensionName: string) => void>();
-
-  private _extensionContext!: ExtensionContext;
-  private _rpc!: ProtoRpcPeer<{ Control: ControlService }>;
-
-  constructor(private readonly opts: ControlExtensionOpts) {}
-
-  async registerExtension(name: string) {
-    await this._rpc.rpc.Control.registerExtension({ name });
-  }
-
-  async onOpen(extensionContext: ExtensionContext): Promise<void> {
-    this._extensionContext = extensionContext;
-
-    // NOTE: Make sure that RPC timeout is greater than the heartbeat timeout.
-    // TODO(dmaretskyi): Allow overwriting the timeout on individual RPC calls?
-    this._rpc = createProtoRpcPeer<ControlRpcBundle, ControlRpcBundle>({
-      requested: {
-        Control: schema.getService('dxos.mesh.teleport.control.ControlService'),
-      },
-      exposed: {
-        Control: schema.getService('dxos.mesh.teleport.control.ControlService'),
-      },
-      handlers: {
-        Control: {
-          registerExtension: async (request) => {
-            this.onExtensionRegistered.call(request.name);
-          },
-          heartbeat: async (request) => {
-            // Ok.
-          },
-        },
-      },
-      port: await extensionContext.createPort('rpc', {
-        contentType: 'application/x-protobuf; messagType="dxos.rpc.Message"',
-      }),
-    });
-
-    await this._rpc.open();
-
-    scheduleTaskInterval(
-      this._ctx,
-      async () => {
-        try {
-          await asyncTimeout(this._rpc.rpc.Control.heartbeat(), this.opts.heartbeatTimeout);
-        } catch (err: any) {
-          this.opts.onTimeout();
-        }
-      },
-      this.opts.heartbeatInterval,
-    );
-  }
-
-  async onClose(err?: Error): Promise<void> {
-    await this._ctx.dispose();
-    await this._rpc.close();
-  }
-
-  async onAbort(err?: Error | undefined): Promise<void> {
-    await this._ctx.dispose();
-    await this._rpc.abort();
-  }
 }
